Remove LBDB and cleanup adjoining parts of LB framework
